Added script that removes underscores from Sphinx output directories.
GitHub does not copy files beginning with an underscore to the directory used for GitHub Pages. This is a problem since Sphinx creates directories named '_images' and '_static' as part of its output. These directories are now renamed to 'images' and 'static', respecively, and the links to files in these directories are adjusted accordingly.