SystemPropertiesDialogHandler no longer allows ProductionSystem properties to exceed ten thousand.
This fixes a bug in ParameterWidgetHandler. Previously, the loop in getCurrents wouldn't terminate if, for example, _maxCurrent was 1e16 and _currentIncrement was 0.1 since 1e16 + 0.1 == 1e16.

I'll think about further lowering the maximum values for maxHeatingCurrent and maxSafeTemperatureSensorVoltage when I refactor the gui.systemprops module.

I've also rewritten the loop in getCurrents to use multiplication rather than a series of additions. As a result, the method will raise an OverflowError rather than entering an endless loop if _maxCurrent somehow manages to become absurdly high.